% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_unreal.R
\name{\%unreal <-\%}
\alias{\%unreal <-\%}
\title{In-place unreal replacers}
\usage{
x \%unreal <-\% replacement
}
\arguments{
\item{x}{a vector or matrix whose unreal values are to be replaced.}

\item{replacement}{the replacement value.}
}
\value{
#' @return
This operator does not return any value:
it is an in-place modifiers, and thus modifies x directly.
The x vector is modified such that all NA, NaN and infinities are replaced with the given replacement value.
}
\description{
In-place modifiers to replace unreal (NA, NaN, Inf, -Inf) elements. \cr
Works on vectors and matrices.
}
\examples{
x <- c(1:9, NA, NaN, Inf)
print(x)
x \%unreal <-\% 0 # same as x[is.na(x)|is.nan(x)|is.infinite(x)] <- 0
print(x)

}
