% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_diagnose.R
\name{import_diagnose}
\alias{import_diagnose}
\title{Check for Mismatches between Loaded and Installed Packages}
\usage{
import_diagnose(lib.loc = .libPaths())
}
\arguments{
\item{lib.loc}{a character vector describing the location of R library trees to search through. \cr \cr}
}
\value{
If no issues are found, returns \code{NULL}. \cr
Otherwise, a data.frame giving the packages where issues have been found. \cr
This data.frame will have the following columns:
\itemize{
\item "package": Character vector of package names
\item "installed_in_lib.loc": logical vector indicating if the package is actually installed in the given \code{lib.loc} paths (\code{TRUE}) or not (\code{FALSE}).
\item "version_loaded": character vector giving the version of the packages as loaded in \link[base]{loadedNamespaces}.
\item "version_installed": character vector giving the version of the packages as installed in \code{lib.loc}; \cr
It will be \code{NA} if the package is not installed in \code{lib.loc}.
\item "versions_equal": logical vector indicating if the loaded and installed versions match. \cr
Gives \code{NA} if the package is not installed in \code{lib.loc}. \cr \cr
}
}
\description{
The \code{import_diagnose()} function
compares the loaded packages
with those installed in the specified \code{lib.loc},
and checks for version and library path mismatches. \cr
Any differences found will be reported in the form of a simple \code{data.frame}. \cr
\cr
}
\examples{

import_diagnose()



}
\seealso{
\link{tinycodet_import}
}
