% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_env.R
\name{import_env}
\alias{import_env}
\title{Environment Specification in the import_ Functions}
\description{
The \code{env} argument in the \code{import_} functions specify
where the functions, exported objects, or alias object will be placed. \cr
The following can be specified for \code{env}:
\itemize{
\item \code{NULL}: If \code{env = NULL},
the objects will be placed in the caller environment.
\item an environment.
\item a string, giving the name of the search path,
as given by \link[base]{search},
to place the objects in. \cr
If multiple search paths have the specified name, an error is returned.
\item a number larger than 1 and smaller than \code{length(search())},
giving the position of the search path to place the objects in. \cr \cr
}

If \code{env} is a string or number, and thus points to a place in the search path,
the search path evironment is not allowed to be any of the following:
\itemize{
\item a package path (their names start with 'package:')
\item a tools path (their names start with 'tools:')
\item the Global environment ('.GlobalEnv')
\item autoloads search path ('Autoloads')
\item a path at position 1 or \code{length(search())} \cr \cr
}

Attempting to use such a path results in an error. \cr
The user can use the \link[=searchenv_add]{searchenv_} functions,
provided by 'tinycodet',
to safely add or remove custom search paths. \cr
\cr
The default value for \code{env} is \code{NULL}. \cr \cr
}
\examples{

search()
searchenv_add("my_ops")
search()

exports <- import_ls("stringi", c("infix", "rp"))
import_from("stringi", exports, env = "my_ops")

foo <- searchenv_get("my_ops")
all(exports \%in\% names(foo))

search()
searchenv_rm("my_ops", which(search() == "my_ops"))
search()
}
\seealso{
\link{tinycodet_import}
}
