% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helper.R
\name{import_helper}
\alias{import_helper}
\alias{help.import}
\title{Helper Functions for the 'tinycodet' Package Import System}
\usage{
help.import(..., i, alias)
}
\arguments{
\item{...}{further arguments to be passed to \link[utils]{help}.}

\item{i}{either one of the following:
\itemize{
\item a function (use back-ticks when the function is an infix/replacement operator). \cr
Examples: \cr
\code{myfun} , \code{`\%operator\%`} , \code{`fun<-`}, \code{myalias.$some_function}. \cr
If a function, the \code{alias} argument is ignored.
\item a string giving the function name or topic (i.e. \code{"myfun"}, \code{"thistopic"}). \cr
If a string, argument \code{alias} must be specified also.
}}

\item{alias}{an object of class \link{tinyimport_alias} as returned by \link{import_as}.}
}
\value{
For \code{help.import()}: \cr
Opens the appropriate help page. \cr
\cr
}
\description{
The \code{help.import()} function
finds the help file for functions or topics,
including exposed functions/operators as well as functions in a package alias object. \cr
\cr
}
\details{
For \code{help.import(...)}: \cr
Do not use the \code{topic} / \code{package} and
\code{i} / \code{alias} argument sets together.
It's either one set or the other. \cr
For example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
import_as(.str ~ stringi)
import_from("magrittr", import_ls("magrittr", "infix"))
help.import(i = .str$stri_sub)
help.import(i = `\%>\%`)
help.import(i = "stri_sub", alias = .str)
help.import(topic = "\%>\%", package = "magrittr")
help.import("\%>\%", package = "magrittr") # same as previous line

}\if{html}{\out{</div>}}
}
\examples{

import_as(.stri ~ stringi)
.stri$stri_join("a", "b")

import_from("stringi", import_ls("stringi", "infix"))
"a" \%s+\% "b"

attr.import(.stri)
}
\seealso{
\link{tinycodet_import}
}
