% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_legacy.R
\name{import_legacy}
\alias{import_legacy}
\alias{import_inops}
\alias{import_LL}
\title{Legacy import_ functions}
\usage{
import_inops(expose, lib.loc = .libPaths(), ...)

import_LL(package, selection, lib.loc = .libPaths())
}
\arguments{
\item{expose, package}{a single string, giving the name of the R-package.}

\item{lib.loc}{a character vector describing the location of R library trees to search through.}

\item{...}{further arguments passed to \link{import_from} or \link{import_ls}}

\item{selection}{a character vector of function names
(both regular functions and infix operators). \cr
Internal functions or re-exported functions are not supported.}
}
\value{
See \link{import_from}.
}
\description{
These functions are deprecated, and will be removed in a future update. \cr \cr
}
\examples{

import_inops("stringi")
import_LL("stringi", "stri_c")



}
\seealso{
\link{tinycodet_import}, \code{\link[=import_from]{import_from()}}, \code{\link[=import_ls]{import_ls()}}
}
