% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ls.R
\name{import_ls}
\alias{import_ls}
\title{List Exported Objects from Package Namespace}
\usage{
import_ls(
  package,
  types = c("reg", "infix", "rp", "nonfun"),
  re_exports = TRUE,
  lib.loc = .libPaths(),
  print = TRUE
)
}
\arguments{
\item{package}{a single string, giving the package name.}

\item{types}{a character vector, specifying the type. \cr
The following types are supported:
\itemize{
\item "reg": regular functions.
\item "infix": infix operators.
\item "rp": replacement operators, including their base functions.
\item "nonfun": non-functions.
}}

\item{re_exports}{\code{TRUE} or \code{FALSE}, indicating if re-exports of \code{package} should be included. \cr
Default is \code{TRUE}, as that is analogous to the behaviour of base R's \link[base]{::} operator.}

\item{lib.loc}{a character vector describing the location of R library trees to search through.}

\item{print}{\code{TRUE} (default) or \code{FALSE},
indicating if the exported objects should be printed to the console as literal code. \cr
Note that this is a \bold{side-effect},
and does not impact the storable return value of \code{import_ls()}. \cr \cr}
}
\value{
A character vector of exported object names defined in the package. \cr
Can be used programmatically in \link[base]{library} or \link{import_from} functions. \cr
I.e.:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Like so:

ls <- import_ls("packagename", "infix")
library(packagename, include.only = ls)

# Or like so:

ls <- import_ls("packagename", "infix")
import_from(packagename, ls = ls)
}\if{html}{\out{</div>}}
}
\description{
Lists exported objects defined (or re-exported) in a package namespace. \cr
Note that \code{import_ls()} necessary loads the package,
but does not attach the package. \cr
\cr
The return value of \code{import_ls()} is for programmatically dynamic code (see section \code{Value}). \cr
The side-effect of \code{import_ls()} is for syntactically readable code (see section \verb{Side Effect}). \cr \cr
}
\section{Side Effect}{

\bold{(if \code{print = TRUE})} \cr
The returned character vector is printed to your console as literal code. \cr
I.e. \code{'c("obj1", "obj2")'}. \cr
One can then copy-paste the printed literal code, for syntactical clarity. \cr
I.e.:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
# Like so:

import_ls("packagename", "infix") # prints literal code to your console
library(packagename, include.only = ...paste printed literal code here...)

# Or like so:

import_ls("packagename", "infix") # prints literal code to your console
import_from("packagename", ls = ...paste printed literal code here...)
}\if{html}{\out{</div>}}
}

\section{Why Listing Functions by Type is Useful}{

One can import a package under an alias using \link{import_as}. \cr
But using infix operators or replacement operators from an alias
requires convoluted code like so: \cr

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
.alias$`\%op\%`(x, y)
.alias$`fun<-`(x, ..., value)

}\if{html}{\out{</div>}}

Instead, it would be easier to just attach (via \link[base]{library})
or expose (via \link{import_from})
such operators so that they can be used on their own. \cr
The \code{import_ls()} function allows the user to get a list of all functions from a certain type,
like "infix operators" or "replacement operators". \cr
The listed functions can then be passed to
\link[base]{library} (to attach them)
or \link{import_from} (to expose them). \cr
\cr
}

\examples{

# Programmatically Dynamic Code ====

ls <- import_ls("stringi", "infix")
import_from("stringi", ls)

if("bit64" \%installed in\% .libPaths()) {
  ls <- import_ls("bit64", "nonfun")
  import_from("bit64", ls)
}


# Syntactically Readable Code ====

import_ls("stringi", "rp") # copy-pasted the printed literal code
import_from(
  "stringi",
  ls = c("stri_datetime_add", "stri_datetime_add<-", "stri_sub", "stri_sub_all", 
         "stri_sub_all<-", "stri_sub<-", "stri_subset", "stri_subset_charclass", 
         "stri_subset_charclass<-", "stri_subset_coll", "stri_subset_coll<-", 
         "stri_subset_fixed", "stri_subset_fixed<-", "stri_subset_regex", 
         "stri_subset_regex<-", "stri_subset<-")
)


}
\seealso{
\link{tinycodet_import}, \link{import_from}
}
