% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace_math.R
\name{\%+ <-\%}
\alias{\%+ <-\%}
\alias{\%- <-\%}
\alias{\%* <-\%}
\alias{\%/ <-\%}
\alias{\%^ <-\%}
\alias{\%rt <-\%}
\alias{\%logb <-\%}
\alias{\%alogb <-\%}
\title{In-place modifying mathematical arithmetic}
\usage{
x \%+ <-\% y

x \%- <-\% y

x \%* <-\% y

x \%/ <-\% y

x \%^ <-\% p

x \%rt <-\% p

x \%logb <-\% b

x \%alogb <-\% b
}
\arguments{
\item{x}{a number or numeric (or 'number-like') vector, matrix, or array.}

\item{y}{a number, or numeric (or 'number-like') vector, matrix, or array of the same length/dimension as \code{x}.
It gives the number to add, substract, multiply by, or divide by.}

\item{p}{a number, or a numeric vector of the same length as \code{x}.
It gives the power to be used.}

\item{b}{a number, or a numeric vector of the same length as \code{x}.
It gives the logarithmic base to be used.}
}
\value{
These operators do not return any value: they are in-place modifiers, and thus modify \code{x} directly.
}
\description{
In-place modifiers for addition, subtraction, multiplication, division, power, root, logarithm, and anti-logarithm. \cr

\code{x \%+ <-\% y } is the same as \code{x <- x + y} \cr
\cr
\code{x \%- <-\% y } is the same as \code{x <- x - y} \cr
\cr
\code{x \%* <-\% y } is the same as \code{x <- x * y} \cr
\cr
\code{x \%/ <-\% y } is the same as \code{x <- x / y} \cr
\cr
\code{x \%^ <-\% p } is the same as \code{x <- x^p} \cr
\cr
\code{x \%rt <-\% p } is the same as \code{x <- x^(1/p)} \cr
\cr
\code{x \%logb <-\% b } is the same as \code{x <- log(x, base=b)} \cr
\cr
\code{x \%alogb <-\% b } is the same as \code{x <- b^x}; if \code{b=exp(1)}, this is the same as \code{x <- exp(x)} \cr
}
\examples{
x <- matrix(rpois(10, 10), ncol=2)
print(x)
x \%+ <-\% 3 # same as x <- x + 3
print(x)

x <- matrix(rpois(10, 10), ncol=2)
print(x)
x \%- <-\% 3 # same as x <- x - 3
print(x)

x <- matrix(rpois(10, 10), ncol=2)
print(x)
x \%* <-\% 3 # same as x <- x * 3
print(x)

x <- matrix(rpois(10, 10), ncol=2)
print(x)
x \%/ <-\% 3 # same as x <- x / 3
print(x)

x <- matrix(rpois(10, 10), ncol=2)
print(x)
x \%^ <-\% 3 # same as x <- x^3
print(x)

x <- matrix(rpois(10, 10), ncol=2)
print(x)
x \%rt <-\% 3 # same as x <- x^(1/3)
print(x)

x <- matrix(rpois(10, 10), ncol=2)
print(x)
x \%logb <-\% 3 # same as x <- log(x, base=3)
print(x)

x <- matrix(rpois(10, 10), ncol=2)
print(x)
x \%alogb <-\% 3 # same as x <- 3^x
print(x)

x <- 3
print(x)
x \%alogb <-\% exp(1) # same as x <- exp(x)
print(x)
exp(3) # notice this is the same.

}
