% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed_in.R
\name{\%installed in\%}
\alias{\%installed in\%}
\title{Check if one or more packages are installed in the specified library}
\usage{
pkgs \%installed in\% lib.loc
}
\arguments{
\item{pkgs}{a single string, or character vector, with the package name(s).}

\item{lib.loc}{character vector specifying library search path
(the location of R library trees to search through).
This is usually \code{.libPaths()}.
See also \link[base]{loadNamespace}.}
}
\value{
Returns a logical vector, where \code{TRUE} indicates a package is installed,
and \code{FALSE} indicates a package is not installed.
\cr
}
\description{
The \code{pkgs \%installed in\% lib.loc} operator
checks if one or more package(s) \code{pkgs} exist(s) in library location \code{lib.loc}. \cr
Now you no longer have to attach a package with \code{require()} simply to check if it exists. \cr
Moreover, this operator makes it syntactically explicit in your code
where you are looking for your R package(s). \cr
\cr
}
\examples{

\dontrun{
pkgs <- c(unlist(tools::package_dependencies("devtools")), "devtools")
pkgs \%installed in\% .libPaths()
}



}
