% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchenv.R
\name{searchenv}
\alias{searchenv}
\alias{searchenv_add}
\alias{searchenv_rm}
\alias{searchenv_get}
\title{Add, Remove, or Access Search Path Environments}
\usage{
searchenv_add(name, pos = 2L)

searchenv_rm(name, pos)

searchenv_get(name, pos)
}
\arguments{
\item{name}{a single string giving the name for the environment in the \link[base]{search} path.}

\item{pos}{a single positive integer giving the position for the environment in the \link[base]{search} path.}
}
\description{
Functions for safely attaching, removing, or accessing environments from \link[base]{search} path. \cr
\cr
\code{searchenv_add()} adds a new, empty environment to the \link[base]{search} path. \cr
\code{searchenv_rm()} removes a user-defined environment from the \link[base]{search} path. \cr
\code{searchenv_get()} returns a user-defined environment from the \link[base]{search} path. \cr \cr
}
\details{
These functions were designed with safety in mind. \cr
They do not allow adding, removing, or accessing search environments like the following:
\itemize{
\item a package path (their names start with 'package:')
\item a tools path (their names start with 'tools:')
\item the Global environment ('.GlobalEnv')
\item autoloads search path ('Autoloads')
\item a path at position 1 or \code{length(search())} \cr \cr
}

Attempting to add a new search path environment whose name already exists gives an error. \cr
Attempting to remove or access a search path environmeent whose name does not exists gives an error. \cr
\cr
}
\examples{

search()
searchenv_add("my_ops")
search()

exports <- import_ls("stringi", c("infix", "rp"))
import_from("stringi", exports, env = "my_ops")

foo <- searchenv_get("my_ops")
all(exports \%in\% names(foo))

search()
searchenv_rm("my_ops", which(search() == "my_ops"))
search()
}
