% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_arithmetic.R
\name{\%s+\%}
\alias{\%s+\%}
\alias{\%s-\%}
\alias{\%s*\%}
\alias{\%s/\%}
\title{String arithmetic}
\usage{
x \%s+\% y

x \%s-\% p

x \%s*\% n

x \%s/\% p
}
\arguments{
\item{x}{a string or character vector.}

\item{y}{a string, or a character vector of the same length as \code{x}.}

\item{p}{a pattern (regular expression), or character vector of regular expressions of the same length as \code{x},
giving the pattern to look for. \cr
See \code{\link{s_pattern_b}}.}

\item{n}{a number, or a numeric vector of the same length as \code{x}.}
}
\value{
The \code{\%s+\%}, \code{\%s-\%}, and \code{\%s*\%} operators
return a character vector of the same length as \code{x}. \cr
The \code{\%s/\%} returns a integer vector of the same length as \code{x}.
}
\description{
String arithmetic operators. \cr
\cr
The \code{x \%s+\% y } operator is equivalent to \code{paste0(x,y)}. \cr
\cr
The \code{x \%s-\% p } operator removes character/pattern defined in \code{p} from \code{x}. \cr
\cr
The \code{x \%s*\% n } operator repeats every element of \code{x} for \code{n} times,
and glues them together. \cr
\cr
The \code{x \%s/\% p } operator counts how often regular expression or character pattern \code{p}
occurs in each element of \code{x}. \cr
\cr
}
\details{
These operators and functions serve as a way to provide straight-forward string arithmetic,
missing from base R. \cr
\cr
}
\examples{
x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
y <- "a"
p <- "a|e|i|o|u" # same as p <- s_pattern_b("a|e|i|o|u", fixed=FALSE, ignore.case=FALSE, perl=FALSE)
n <- c(3, 2)

x \%s+\% y # =paste0(x,y)
x \%s-\% p # remove all vowels from x
x \%s*\% n
x \%s/\% p # count how often vowels appear in each string of vector x.


#############################################################################


x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
y <- "a"
# pattern with ignore.case=TRUE:
p <- s_pattern_b("A|E|I|O|U", fixed=FALSE, ignore.case=TRUE, perl=FALSE)
n <- c(2, 3)

x \%s+\% y # =paste0(x,y)
x \%s-\% p # remove all vowels from x
x \%s*\% n
x \%s/\% p # count how often vowels appears in each string of vector x.


#############################################################################


p <- s_pattern_b("\\\\v+", perl=TRUE) # perl expression; only works with perl=TRUE
x <- "line1 \n line2"
print(x)
x \%s-\% p # remove vertical line

}
