% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_subset_ops.R
\name{str_subset_ops}
\alias{str_subset_ops}
\alias{\%ss\%}
\alias{\%sget\%}
\alias{\%strim\%}
\title{String subsetting operators}
\usage{
x \%ss\% s

x \%sget\% ss

x \%strim\% ss
}
\arguments{
\item{x}{a string or character vector.}

\item{s}{a numeric vector giving the subset indices.}

\item{ss}{a vector of length 2, or a matrix with 2 columns with \code{nrow(ss)==length(x)}.
The object \code{ss} should consist entirely of non-negative integers
(thus 0, 1, 2, etc. are valid, but -1, -2, -3 etc are not valid).
The first element/column of \code{ss}
gives the number of characters counting from the left side to be extracted/removed from \code{x}.
The second element/column of \code{ss}
gives the number of characters counting from the right side to be extracted/removed from \code{x}.}
}
\value{
The \code{\%ss\%} operator always returns a vector or matrix,
where each element is a single character. \cr
\cr
}
\description{
String subsetting operators. \cr
\cr
The \code{x \%ss\% s } operator
allows indexing a single string as-if it is an iterable object. \cr
\cr
The \code{x \%sget\% ss } operator
gives a certain number of the first and last characters of \code{x}. \cr
\cr
The \code{x \%strim\% ss } operator
removes a certain number of the first and last characters of \code{x}. \cr
\cr
}
\details{
These operators serve as a way to provide straight-forward string sub-setting. \cr
\cr
}
\examples{

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
ss <- c(2,3)
x \%sget\% ss

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
ss <- c(1,0)
x \%sget\% ss

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
ss <- c(2,3)
x \%strim\% ss

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
ss <- c(1,0)
x \%strim\% ss

"hello" \%ss\% 5:1







}
