% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_truth.R
\name{\%s{}\%}
\alias{\%s{}\%}
\alias{\%s!{}\%}
\title{String detection operators}
\usage{
x \%s{}\% p

x \%s!{}\% p
}
\arguments{
\item{x}{a string or character vector.}

\item{p}{either a list with \code{stringi} arguments (see \link{stri_rgx}),
or else a character vector of the same length as \code{x} with regular expressions.}
}
\value{
The \code{x \%s{}\% p} and \code{x \%s!{}\% p} operators
return logical vectors. \cr
}
\description{
The \code{x \%s{}\% p} operator
checks for every string in character vector \code{x} if
the pattern defined in \code{p} is present. \cr
\cr
The \code{x \%s!{}\% p} operator
checks for every string in character vector \code{x} if
the pattern defined in \code{p} is NOT present. \cr
}
\examples{

# simple pattern ====

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
x \%s{}\% "a"
x \%s!{}\% "a"
which(x \%s{}\% "a")
which(x \%s!{}\% "a")
x[x \%s{}\% "a"]
x[x \%s!{}\% "a"]

x \%s{}\% "1"
x \%s!{}\% "1"
which(x \%s{}\% "1")
which(x \%s!{}\% "1")
x[x \%s{}\% "1"]
x[x \%s!{}\% "1"]


#############################################################################

# ignore case pattern ====


x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
p <- list(regex = c("A", "A"), case_insensitive=TRUE)
x \%s{}\% p
x \%s!{}\% p
which(x \%s{}\% p)
which(x \%s!{}\% p)
x[x \%s{}\% p]
x[x \%s!{}\% p]


#############################################################################

# multi-character pattern ====

x <- c(paste0(letters[1:13], collapse=""), paste0(letters[14:26], collapse=""))
print(x)
p <- list(regex = rep("AB", 2), case_insensitive=TRUE)
x \%s{}\% p
x \%s!{}\% p
which(x \%s{}\% p)
which(x \%s!{}\% p)
x[x \%s{}\% p]
x[x \%s!{}\% p]



}
\seealso{
\code{\link[=tinyoperations_stringi]{tinyoperations_stringi()}}
}
