% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycodet_dry.R
\name{tinycodet_dry}
\alias{tinycodet_dry}
\title{The tinycodet "DRY" functionality}
\usage{
tinycodet_dry()
}
\description{
"Don't Repeat Yourself", sometimes abbreviated as "DRY",
is the coding principle not to write unnecessarily repetitive code.
To help you in that effort, the \code{tinycodet} R package
introduces a few functions:
\itemize{
\item The \link{transform_if} function
\item The \link[=\%unreal =\%]{subset_if operators and the in-place unreal modifier operator}.
\item The \link[=\%:=\%]{generalized in-place (mathematical) modification operator}.
}
}
\seealso{
\code{\link[=tinycodet_help]{tinycodet_help()}}
}
