% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycodet_help.R
\name{tinycodet_help}
\alias{tinycodet_help}
\title{The tinycodet help page}
\usage{
tinycodet_help()
}
\description{
Welcome to the \code{tinycodet} introduction help page! \cr
\cr
The 'tinycodet' R-package is a tiny R package that adds a few functions to help in your coding etiquette. \cr
It primarily focuses on 4 things: \cr
\cr
(1) Safer decimal (in)equality testing, safer atomic conversions, and other functions for safer coding; \cr
see \link{tinycodet_safer}. \cr
\cr
(2) A new package import system,
that attempts to combine the benefits of using a package without attaching,
with the benefits of attaching a package; \cr
see \link{tinycodet_import} \cr
\cr
(3) Extending the string manipulation capabilities of the 'stringi' R package; \cr
see \link{tinycodet_strings}. \cr
\cr
(4) Reducing repetitive code; \cr
see \link{tinycodet_dry}. \cr
\cr
And some miscellaneous functionality; see \link{tinycodet_misc}. \cr
\cr

The \code{tinycodet} R-package has only one dependency,
namely 'stringi'. No other dependencies, thus avoiding "dependency hell".
Most functions in this R-package are fully vectorized
and optimized, and have been well documented.
\cr
}
\references{
The badges shown in the documentation of this R package were made using the services of: \url{https://shields.io/}
}
