% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycodet_safer.R
\name{tinycodet_safer}
\alias{tinycodet_safer}
\title{The tinycodet "safer" functionality}
\usage{
tinycodet_safer()
}
\description{
To help make your code safer, the \code{tinycodet} R package
introduces a few functions:
\itemize{
\item \link[=\%d==\%]{Safer decimal (in)equality testing}.
\item \link[=as_int]{Atomic type casting without stripping attributes}.
\item The \link{lock_TF} function to
set and lock \code{T} and \code{F} to \code{TRUE} and \code{FALSE}.
\item The \link{\%<-c\%} operator to assign locked constants.
}
}
\seealso{
\code{\link[=tinycodet_help]{tinycodet_help()}}
}
