% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycodet_strings.R
\name{tinycodet_strings}
\alias{tinycodet_strings}
\title{The tinycodet expansion of the 'stringi' R package}
\usage{
tinycodet_strings()
}
\description{
The \code{tinycodet} R package adds some functions and operators
to extend the functionality of the \code{stringi} R package:
\itemize{
\item Infix operators for \link[=\%s-\%]{string arithmetic}. \cr
\item Infix operators for \link[=\%ss\%]{string sub-setting}. \cr
\item \link[=\%row~\%]{Infix operators for row- and column-wise re-ordering of matrices}. \cr
\item The \code{tinycodet} package adds additional
\code{stringi} functions, namely \link{stri_locate_ith}, and
\link{stri_join_mat} (and aliases). These functions use the same naming and argument convention as the rest of
the \code{stringi} functions, thus keeping your code consistent. \cr
\item The \link[=strcut_loc]{strcut_-functions}.
\item Most \code{stringi} pattern expressions options
are available for the string-pattern-related functions, when appropriate. \cr
\item This R package has only one dependency: \code{stringi}.
No other dependencies, as to avoid \code{"dependency hell"}. \cr
\item Although the functions are written in R,
they have been aggressively optimized to be in the same order of speed
as the other \code{stringi} functions.
}
}
\references{
Gagolewski M., stringi: Fast and portable character string processing in R, Journal of Statistical Software 103(2), 2022, 1–59, doi:10.18637/jss.v103.i02
}
\seealso{
\code{\link[=tinycodet_help]{tinycodet_help()}}
}
