% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyoperations_import.R
\name{tinyoperations_import}
\alias{tinyoperations_import}
\title{The tinyoperations import system}
\usage{
tinyoperations_import()
}
\description{
The \code{tinyoperations} R package introduces a new package import system.
This system attempts to combine the benefits of aliasing a package,
with the benefits of attaching a package. \cr
\cr
The main part of the import system is implemented in the 3 \code{import_} - functions:
\itemize{
\item \link{import_as}:
Allow a main package + its foreign exports + its dependencies + its enhances + its extensions
to be loaded under a single alias.
\item \link{import_inops}:
Allow exposing infix operators to the current environment.
\item \link{import_data}:
Allow assigning a data set from a package directly to a variable.
}

The above functionality is also extended to work on sourced modules,
see \link[=source_inops]{source_module}. \cr
\cr
There are also some additional helper functions for the package import system,
see \link[=pkg_get_deps]{pkgs}. \cr
\cr
Please refer to the Read-Me file on the GitHub main page of this page for more information. \cr
See: \url{https://github.com/tony-aw/tinyoperations}. \cr
}
\seealso{
\link{tinyoperations_help}
}
