% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyoperations_misc.R
\name{tinyoperations_misc}
\alias{tinyoperations_misc}
\title{The tinyoperations miscellaneous functionality}
\usage{
tinyoperations_misc()
}
\description{
Some additional functions provided by the \code{tinyoperations} R package:
\itemize{
\item \link[=\%xor\%]{Infix logical operators} for exclusive-or, not-and, not-in, number-type, and string-type. \cr
\item \link[=lock_TF]{restrict re-assignment of "T" and "F"}.
\item \link[=\%<-c\%]{create unchangeable CONSTANT}.
\item \link[=\%row~\%]{Infix operators for row- and column-wise re-ordering of matrices}.
}

Please refer to the Read-Me file on the GitHub main page of this page for more information. \cr
See: \url{https://github.com/tony-aw/tinyoperations}. \cr
}
\seealso{
\code{\link[=tinyoperations_help]{tinyoperations_help()}}
}
