% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyoperations_safer.R
\name{tinyoperations_safer}
\alias{tinyoperations_safer}
\title{The tinyoperations "safer" functionality}
\usage{
tinyoperations_safer()
}
\description{
To help make your code safer, the \code{tinyoperations} R package
introduces a few functions:
\itemize{
\item \link[=\%d==\%]{Safer decimal (in)equality testing}.
\item \link[=as_int]{Atomic type casting without stripping attributes}.
\item The \link{lock_TF} function to
set and lock \code{T} and \code{F} to \code{TRUE} and \code{FALSE}.
\item The \link{\%<-c\%} operator to assign locked constants.
}

Please refer to the Read-Me file on the GitHub main page of this page for more information. \cr
See: \url{https://github.com/tony-aw/tinyoperations}. \cr
}
\seealso{
\code{\link[=tinyoperations_help]{tinyoperations_help()}}
}
