% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_if.R
\name{transform_if}
\alias{transform_if}
\title{The transform_if function}
\usage{
transform_if(x, cond, yes = function(x) x, no = function(x) x, other = NA)
}
\arguments{
\item{x}{a vector, matrix, or array.}

\item{cond}{either a \code{logical} vector,
or a (possibly anonymous) function that returns a \code{logical} vector. \cr
In both cases the logical vector in question must be of the same length as \code{x}. \cr
For example: \code{\(x)x>0}. \cr}

\item{yes}{the (possibly anonymous) transformation function to use
when \code{cond(x)==TRUE}. \cr
Alternatively, one can also supply an atomic scalar to fill in
when \code{cond(x)==TRUE}. \cr
If this argument is not specified, \code{yes} defaults to \code{function(x)x}.}

\item{no}{the (possibly anonymous) transformation function to use
when \code{cond(x)==FALSE}. \cr
Alternatively, one can also supply an atomic scalar to fill in
when \code{cond(x)==FALSE}. \cr
If this argument is not specified, \code{no} defaults to \code{function(x)x}.}

\item{other}{the (possibly anonymous) transformation function to use
when \code{cond(x)} returns \code{NA}. \cr
Alternatively, one can also supply an atomic scalar. \cr
If this argument is not specified, argument \code{other} defaults to \code{NA}. \cr
Note that function  \code{other(x)} is run or scalar \code{other} is returned
when \code{cond(x)} or \code{cond} is \code{NA},
not necessarily when \code{x} itself is \code{NA}.}
}
\value{
The transformed vector, matrix, or array (attributes are conserved).
}
\description{
The \code{transform_if()} function transforms an object \code{x},
based on the logical result (\code{TRUE, FALSE, NA})
of condition function \code{cond(x)} or logical vector \code{cond},
such that: \cr

\itemize{
\item For every value where \code{cond(x)==TRUE} / \code{cond==TRUE},
function \code{yes(x)} is run or scalar \code{yes} is returned.
\item For every value where \code{cond(x)==FALSE} / \code{cond==FALSE},
function \code{no(x)} is run or scalar \code{no} is returned.
\item For every value where \code{cond(x)==NA} / \code{cond==NA},
function \code{other(x)} is run or scalar \code{other} is returned. \cr
}
}
\details{
Be careful with coercion! For example the following code: \cr
\code{x <- c("a", "b")} \cr
\code{transform_if(x, \(x)x=="a", as.numeric, as.logical)} \cr
returns: \cr
\code{[1] NA NA} \cr
due to the same character vector being given 2 incompatible classes. \cr
\cr
}
\examples{
x <- c(-10:9, NA, NA)
object <- matrix(x, ncol=2)
attr(object, "helloworld") <- "helloworld"
print(object)
y <- 0
z <- 1000
object |> transform_if(\(x)x>y, log, \(x)x^2, \(x)-z)
object |> transform_if(object > y, log, \(x)x^2, -z)

}
\seealso{
\code{\link[=tinyoperations_dry]{tinyoperations_dry()}}
}
