% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{env_c}
\alias{env_c}
\title{Utility operators and functions}
\usage{
env_c(...)
}
\arguments{
\item{...}{environments to be concatenated.}
}
\value{
A merged environment
}
\description{
The \code{env_c} function concatenates multiple environments into a single environment. \cr
This very handy, for example,
when one wants to load multiple R packages under a single alias.
I.e.: \cr
\code{fv <- env_c(loadNamespace("tidytable"), loadNamespace("data.table"), loadNamespace("collapse"))} \cr
\cr
}
