% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xxx_atomic_typecast.R
\name{xxx_atomic_typecast}
\alias{xxx_atomic_typecast}
\alias{as_bool}
\alias{as_int}
\alias{as_dbl}
\alias{as_chr}
\alias{as_cplx}
\alias{as_raw}
\title{DEPRECATED: Atomic Type Casting With Names and Dimensions Preserved}
\usage{
as_bool(x, ...)

as_int(x, ...)

as_dbl(x, ...)

as_chr(x, ...)

as_cplx(x, ...)

as_raw(x, ...)
}
\arguments{
\item{x}{vector, matrix, array
(or a similar object where all elements share the same type).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The converted object. \cr \cr
}
\description{
THESE FUNCTIONS ARE DEPRECATED, AND WILL BE REMOVED IN A FUTURE VERSION. \cr
Please use the functions of the same name from the 'broadcast' package instead. \cr
}
\examples{

# matrix example ====
x <- matrix(sample(-1:28), ncol = 5)
colnames(x) <- month.name[1:5]
rownames(x) <- month.abb[1:6]
names(x) <- c(letters[1:20], LETTERS[1:10])
print(x)

as_bool(x)
as_int(x)
as_dbl(x)
as_chr(x)
as_cplx(x)
as_raw(x)


################################################################################

# factor example ====
x <- factor(month.abb, levels = month.abb)
names(x) <- month.name
print(x)

as_bool(as_int(x) > 6)
as_int(x)
as_dbl(x)
as_chr(x)
as_cplx(x)
as_raw(x)
}
