% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycodet_safer.R
\name{aaa1_tinycodet_safer}
\alias{aaa1_tinycodet_safer}
\alias{tinycodet_safer}
\title{Overview of the 'tinycodet' "Safer" Functionality}
\description{
To help make your code safer, the 'tinycodet' R-package
introduces a few functions:
\itemize{
\item \link[=\%d==\%]{Safer decimal (in)equality testing}.
\item \link[=as_int]{Atomic type casting without stripping attributes}.
\item The \link{lock_TF} function to
set and lock \code{T} and \code{F} to \code{TRUE} and \code{FALSE}.
\item The \link{\%<-c\%} operator to assign locked constants.
}
}
\examples{
x <- c(0.3, 0.6, 0.7)
y <- c(0.1*3, 0.1*6, 0.1*7)
x == y # gives FALSE, but should be TRUE
x \%d==\% y # here it's done correctly



}
\seealso{
\link{tinycodet_help}
}
