% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa5_tinycodet_misc.R
\name{aaa5_tinycodet_misc}
\alias{aaa5_tinycodet_misc}
\alias{tinycodet_misc}
\title{Overview of the 'tinycodet' Miscellaneous Functionality}
\description{
Some additional functions provided by the 'tinycodet' R-package:
\itemize{
\item \link[=\%xor\%]{Infix logical operators} for exclusive-or, not-and, not-in, number-type, and string-type.
\item \link[=\%row~\%]{Infix operators for row- and column-wise re-ordering of matrices}.
\item \link[=report_inops]{Report infix operators present in the current environment, or a specified environment.}
\item \link{source_selection} to source only selected objects. \cr \cr
}
}
\seealso{
\code{\link[=tinycodet_help]{tinycodet_help()}}
}
