% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\alias{\%m import <-\%}
\alias{import_data}
\alias{import_lsf}
\title{Package import management operator and functions}
\usage{
alias \%m import <-\% pkgs

import_data(dataname, package)

import_lsf(package, type)
}
\arguments{
\item{alias}{a variable name (unquoted),
giving the (not yet existing) object
where the package(s) are to be assigned to.}

\item{pkgs}{a character vector with the package name(s). \cr
NOTE: The order matters! If 2 packages share objects with the same name,
the package named last will overwrite the earlier named package.}

\item{dataname}{a single string, giving the name of the dataset.}

\item{package}{a single string, giving the name of the package.}

\item{type}{The type of functions to list. Possibilities: \cr
\code{"inops"} or \code{"operators"}: Only infix operators (functions surrounded by percentage signs). \cr
\code{"regfuns"}: Only regular functions (thus excluding infix operators). \cr
\code{"all"}: All functions, both regular and infix operators. \cr}
}
\value{
For \code{\%m import <-\%}: \cr
The variable named in the \code{alias} argument will be created
(if it did not already exist),
and it will contain the (merged) package environment. \cr
\cr
For \code{import_data()}: \cr
Returns the data directly.
Thus, one can assign the data like so: \code{mydata <- import_data(...)}. \cr
\cr
For \code{import_lsf()}: \cr
A vector of function and/or operator names. \cr
This vector can be directly used in the \code{include.only} argument of the
\code{library()} function. \cr
}
\description{
The \code{alias \%m import <-\% pkgs} operator
imports the namespaces of an R package
(or a small set of R packages that "belong" to each other)
under the same alias. \cr
\cr
The \code{import_lsf(package, ...)} function gets a list of exported functions from a package. \cr
\cr
The \code{import_data(dataname, package)} function gets a specified data set from a package. \cr
Unlike \code{utils::data()}, the \code{import_data()} function returns the dataset directly,
and allows assigning the dataset like so: \cr
\code{mydata <- import_data(...)}. \cr
\cr
}
\details{
The \code{alias \%m import <-\% pkgs} command is essentially the same as \cr
\code{alias <- loadNamespace("packagename")} \cr
except the \code{alias \%m import <-\% pkgs} operator
allows assigning multiple packages to the same alias,
and this operator does not import internal functions
(i.e. internal functions are kept internal, as they should). \cr
\cr
The \code{alias \%m import <-\% pkgs} operator will tell the user
about conflicting objects. It will also inform the user when importing
a package that consists mostly of infix operators. \cr
\cr
Note: the user should not use the \code{alias \%m import <-\% pkgs} operator
unless the user knows what he/she is doing. \cr
The operator will give a warning when more than 3 packages being imported into the same alias. \cr
\cr
}
\examples{

\dontrun{
fv \%m import <-\% c("data.table", "collapse", "tidytable")
library(data.table, include.only = import_lsf("data.table", type="inops"))
library(collapse, include.only = import_lsf("collapse", type="inops"))
d <- import_data("chicago", "gamair")
}



}
