% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_as.R
\name{import_as}
\alias{import_as}
\title{Import R-package (and Minimal Dependencies) Under an Alias}
\usage{
import_as(
  main,
  re_exports = TRUE,
  deps = NULL,
  lib.loc = .libPaths(),
  env = NULL
)
}
\arguments{
\item{main}{a 2-sided formula
(or a string that evaluates as a 2-sided formula),
where the left-hand side gives the alias,
and the right-hand side gives the main package to import under the given alias. \cr
Starting the alias name with a dot will hide it from \code{ls()},
preventing accidental removal. \cr
For example : \cr
\code{.alias ~ packagename} \cr
Core R (i.e. "base", "stats", etc.) is not allowed for the main package.}

\item{re_exports}{\code{TRUE} or \code{FALSE}.
\itemize{
\item If \code{re_exports = TRUE} the re-exports from the main package
(including those exported from Core R)
are added to the alias together with the main package. \cr
This is the default,
as it is analogous to the behaviour of base R's \link[base]{::} operator. \cr
\item If \code{re_exports = FALSE},
these re-exports are not added together with the main package. \cr
The user can still import the packages under the alias from which the re-exported functions came from,
by specifying them in the \code{dependencies} argument.
}}

\item{deps}{an optional character vector,
giving the names of the dependencies of the
main package to be imported also under the alias. \cr
Only dependencies that appear in \link{pkg_get_deps_minimal} are allowed. \cr
Defaults to \code{NULL}, which means no dependencies are imported under the alias. \cr
Core R (i.e. "base", "stats", etc.) is not allowed.}

\item{lib.loc}{a character vector describing the location of R library trees to search through.}

\item{env}{see \link{import_env}. \cr \cr}
}
\value{
A locked environment object, similar to the output of \link[base]{loadNamespace},
with the name as specified in the \code{alias} argument,
will be created. \cr
This object, of class \link{tinyimport_alias},
will contain the exported functions from the specified package(s). \cr
The alias object will be placed in the specified environment. \cr
For its usage, see \link{tinyimport_alias}. \cr
\cr
Note the following:
\itemize{
\item No more than 5 packages
(ignoring re-exports)
are allowed to be imported under a single alias.
\item Packages are imported in the following order: \cr
First the dependencies  in the order they are specified in \code{deps},
and then the main package, and then its re-exports (if \code{re_exports = TRUE}). \cr
Thus main package will always overwrite the dependencies in case of conflicting names. \cr
}
}
\description{
The \code{import_as()} function
imports the namespace of an R-package,
and optionally also its direct minimal dependencies,
all under the same alias.
The specified alias,
containing the exported functions from the specified packages,
will be placed in the specified environment. \cr
}
\section{Why Aliasing A Package with its Dependencies is Useful}{

To use an R-package with its dependencies,
whilst avoiding the disadvantages of attaching a package (see \link{tinycodet_import}),
one would traditionally use the \link[base]{::} operator like so: \cr

\if{html}{\out{<div class="sourceCode r">}}\preformatted{main_package::some_function1()
dependency1::some_function2()
}\if{html}{\out{</div>}}

This becomes cumbersome as more packages are needed and/or
as the package name(s) become longer. \cr
The \code{import_as()} function avoids this issue
by allowing multiple \bold{related} packages to be imported under a single alias,
allowing one to code like this:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{import_as(.alias ~ main_package, deps = "dependency1")
.alias$some_function1()
.alias$some_function2()
}\if{html}{\out{</div>}}

Thus importing a package, or multiple directly related packages, under a single alias,
which \code{import_as()} provides, avoids the above issues.
Importing a package under an alias is referred to as "aliasing" a package. \cr \cr
}

\examples{

import_as(.stri ~ stringi)
.stri$stri_join("a", "b")

import_from("stringi", import_ls("stringi", "infix"))
"a" \%s+\% "b"

attr.import(.stri)
}
\seealso{
\link{tinycodet_import}
}
