% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from.R
\name{import_from}
\alias{import_from}
\title{Expose Exported Objects From Package Namespace in an Environment}
\usage{
import_from(
  package,
  ls,
  re_exports = TRUE,
  lock = FALSE,
  prefix = NULL,
  lib.loc = .libPaths(),
  env = NULL
)
}
\arguments{
\item{package}{a single string, giving the package name.}

\item{ls}{a character vector giving the names of exported objects to expose. \cr
To expose all objects of a certain type,
like all (infix/replacement) operators or all non-functions,
the \link{import_ls} function can be used to get those object names.}

\item{re_exports}{\code{TRUE} or \code{FALSE}, indicating if re-exports of \code{package} should be included. \cr
Default is \code{TRUE}, as that is analogous to the behaviour of base R's \link[base]{::} operator.}

\item{lock}{\code{TRUE} or \code{FALSE}, indicating if the exported objects should be locked
(see \link[base]{lockBinding}).}

\item{prefix}{OPTIONAL: either \code{NULL} or a single string, giving the prefix to add to the functions, to avoid possible conflicts. \cr
If \code{NULL} no prefix is added; this is the default. \cr
Otherwise, a prefix is added to the exported object names. \cr
Note that \bold{no} prefix is added to infix operators, primitive functions, and non-functions (like constants). \cr
The prefix cannot end with a single dot, as that might interfere with S3 methods.}

\item{lib.loc}{a character vector describing the location of R library trees to search through.}

\item{env}{see \link{import_env}. \cr \cr}
}
\value{
The objects specified in the given package will be placed & locked
in the specified environment. \cr
\cr
}
\description{
\code{import_from()}
exposes exported objects from a package to the specified environment. \cr \cr
}
\examples{

import_from("stringi", import_ls("stringi", "infix"))
import_from("stringi", import_ls("stringi", "rp"))



}
\seealso{
\link{tinycodet_import}, \code{\link[=import_ls]{import_ls()}}
}
