% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace.R
\name{inplace}
\alias{inplace}
\alias{\%:=\%}
\title{Generalized in-place modifier}
\usage{
x \%:=\% f
}
\arguments{
\item{x}{a variable.}

\item{f}{a (possibly anonymous) function to be applied in-place on \code{x}.
The function must take one argument only.}
}
\value{
This operator does not return any value: \cr
It is an in-place modifier, and thus modifies the object directly.
}
\description{
Generalized in-place modifier. \cr
\cr
The \code{x \%:=\% f} operator allows performing
in-place modification of some object \code{x}
with a function \code{f}. \cr
\cr
For example this: \cr
\code{mtcars$mpg[mtcars$cyl>6] <- mtcars$mpg[mtcars$cyl>6]^2} \cr
Can now be re-written as: \cr
\code{mtcars$mpg[mtcars$cyl>6] \%:=\% \(x)x^2} \cr
\cr
}
\examples{
set.seed(1)
object <- matrix(rpois(10, 10), ncol=2)
print(object)
y <- 3
object \%:=\% \(x) x+y # same as object <- object + y
print(object)

}
\seealso{
\code{\link[=tinycodet_dry]{tinycodet_dry()}}
}
