% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{lock_TF}
\alias{\%<-c\%}
\title{Miscellaneous functions to help your coding etiquette}
\usage{
lock_TF()

X \%<-c\% A
}
\arguments{
\item{X}{a syntactically valid unquoted name of the object to be created.}

\item{A}{any kind of object to be assigned to \code{X}.}
}
\value{
For \code{lock_TF()}: \cr
Two \code{constants}, namely \code{T} and \code{F},
set to \code{TRUE} and \code{FALSE} respectively,
are created in the current environment,
and locked.
Removing the created \code{T} and \code{F} objects allows re-assignment again. \cr
\cr
For \code{X \%<-c\% A}: \cr
The object \code{X} containing \code{A} is created in the current environment,
and this object cannot be changed. It can only be accessed or removed.
}
\description{
The \code{lock_TF()} function
locks the \code{T} and \code{F} values and sets them to \code{TRUE} and \code{FALSE},
to prevent the user from re-assigning them. \cr
Removing the created \code{T} and \code{F} objects
allows re-assignment again. \cr
\cr
The \code{X \%<-c\% A} operator creates a \code{constant} \code{X}
and assigns \code{A} to it. \cr
Constants cannot be changed, only accessed or removed.
So if you have a piece of code that requires some unchangeable \code{constant},
use this operator to create said \code{constant}. \cr
Removing constant \code{X} also removes its binding lock.
Thus to change a \code{constant}, simply remove it and re-create it. \cr
}
\examples{

lock_TF()
X \%<-c\% data.frame(x=3, y=2) # this data.frame cannot be changed. Only accessed or removed.
X[1, ,drop=FALSE]


}
\seealso{
\code{\link[=tinyoperations_misc]{tinyoperations_misc()}}
}
