% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{stricter_TrueFalse}
\title{Miscellaneous functions to help your coding etiquette}
\usage{
stricter_TrueFalse()
}
\value{
Two objects: \code{T} and \code{F}, both set to \code{NULL}.
Removing the created \code{T} and \code{F} objects will restore the behaviour. \cr
}
\description{
The \code{stricter_TrueFalse} function re-assigns the \code{T} and \code{F} values to
\code{NULL}, forcing the user to use \code{TRUE} and \code{FALSE}.
Removing the created \code{T} and \code{F} objects will restore the behaviour. \cr
The \code{}
}
\examples{

stricter_TrueFalse()



}
