% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{misc}
\alias{misc}
\alias{stricter_TrueFalse}
\alias{\%<-c\%}
\title{Miscellaneous functions to help your coding etiquette}
\usage{
stricter_TrueFalse()

X \%<-c\% A
}
\arguments{
\item{X}{a syntactically valid unquoted name of the object to be created.}

\item{A}{any kind of object to be assigned to \code{X}.}
}
\value{
For \code{stricter_TrueFalse()}: \cr
Two \code{CONSTANTS}, namely \code{T} and \code{F},
set to \code{TRUE} and \code{FALSE} respectively,
and locked.
Removing the created \code{T} and \code{F} objects allows re-assignment again. \cr
\cr
For \code{X \%<-c\% A}: \cr
The object \code{X} containing \code{A} is created in the current environment,
and this object cannot be changed. It can only be accessed or removed.
}
\description{
The \code{stricter_TrueFalse()} function
locks the \code{T} and \code{F} values,
to prevent the user from re-assigning them. \cr
Removing the created \code{T} and \code{F} objects
allows re-assignment again. \cr
\cr
The \code{X \%<-c\% A} operator creates a \code{CONSTANT} \code{X}
and assigns \code{A} to it. \cr
Constants cannot be changed, only accessed or removed.
So if you have a piece of code that absolutely requires some \code{CONSTANT},
use this operator to create said \code{CONSTANT}. \cr
Removing object \code{X} also removes its binding lock.
Thus to change a \code{CONSTANT}, simply remove it and re-create it. \cr
}
\examples{

stricter_TrueFalse()
X \%<-c\% data.frame(x=3, y=2) # this data.frame cannot be changed. Only accessed or removed.
X[1, ,drop=FALSE]


}
