% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs.R
\name{pkgs_get_deps}
\alias{pkgs_get_deps}
\alias{\%installed in\%}
\title{Miscellaneous package functions}
\usage{
pkgs_get_deps(
  package,
  lib.loc = .libPaths(),
  deps_type = c("Depends", "Imports", "LinkingTo"),
  base = FALSE,
  recom = FALSE
)

pkgs \%installed in\% lib.loc
}
\arguments{
\item{package}{a single string giving the package name}

\item{lib.loc}{character vector specifying library search path
(the location of R library trees to search through).
This is usually \code{.libPaths()}.
See also \link[base]{loadNamespace}.}

\item{deps_type}{a character vector, giving the dependency types to be used. \cr
Defaults to \code{c("Depends", "Imports", "LinkingTo")}.}

\item{base}{logical,
indicating whether base/core R packages should be included (\code{TRUE}),
or not included (\code{FALSE}; the default).}

\item{recom}{logical,
indicating whether the pre-installed "recommended" R packages should be included (\code{TRUE}),
or not included (\code{FALSE}; the default).
Note that only the recommended R packages actually installed in your system are taken into consideration.}

\item{pkgs}{a single string, or character vector, with the package name(s).}
}
\value{
For  \code{pkgs \%installed in\% lib.loc}:
Returns a logical vector, where \code{TRUE} indicates a package is installed,
and \code{FALSE} indicates a package is not installed. \cr
\cr
For \code{pkgs_get_deps()}:
A character vector of dependencies.
}
\description{
The \code{pkgs \%installed in\% lib.loc} operator
checks if one or more package(s) \code{pkgs} exist(s) in library location \code{lib.loc}. \cr
Now you no longer have to attach a package with \code{require()} simply to check if it exists. \cr
Moreover, this operator makes it syntactically explicit in your code
where you are looking for your R package(s). \cr
\cr
The \code{pkgs_get_deps()} function gets the dependencies of a package
from the Description file. It works on non-CRAN packages also.
}
\examples{

\dontrun{
pkgs <- c(unlist(tools::package_dependencies("devtools")), "devtools")
pkgs \%installed in\% .libPaths()
}



}
\references{
https://stackoverflow.com/questions/30223957/elegantly-extract-r-package-dependencies-of-a-package-not-listed-on-cran
}
