% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_arithmetic.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%s+\%}
\alias{\%s*\%}
\alias{\%s$\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{stringi}{\code{\link[stringi:operator_dollar]{\%s$\%}}, \code{\link[stringi:stri_dup]{\%s*\%}}, \code{\link[stringi:operator_add]{\%s+\%}}}
}}

