% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safer_match.R
\name{safer_partialmatch}
\alias{safer_partialmatch}
\title{Set Safer Dollar, Arguments, and Attribute Matching}
\usage{
safer_partialmatch()
}
\value{
Sets the options. Returns nothing. \cr
}
\description{
The \code{safer_partialmatch()} function simply calls the following:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(
   warnPartialMatchDollar = TRUE,
   warnPartialMatchArgs = TRUE,
   warnPartialMatchAttr = TRUE
 )

}\if{html}{\out{</div>}}

Thus it forces 'R' to give a warning when partial matching occurs when using
the dollar (\link{$}) operator,
or when other forms of partial matching occurs. \cr
The \code{safer_partialmatch()} function
is intended for when running R interactively
(see \link[base]{interactive}). \cr \cr
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
interactive()


safer_partialmatch()
data(iris)
head(iris)
iris$Sepal.Length <- iris$Sepal.Length^2
head(iris)
\dontshow{\}) # examplesIf}
}
\seealso{
\link{tinycodet_safer}
}
