% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycodet_strings.R
\name{tinycodet_strings}
\alias{tinycodet_strings}
\title{Overview of the tinycodet Extension of 'stringi'}
\usage{
tinycodet_strings()
}
\description{
The 'tinycodet' R-package adds some functions and operators
to extend the functionality of the 'stringi' R-package:
\itemize{
\item Infix operators for \link[=\%s-\%]{string arithmetic}. \cr
\item Infix operators for \link[=\%sget\%]{string sub-setting}. \cr
\item Infix operators for \link[=\%s\{\}\%]{detecting patterns}. \cr
\item Find  \eqn{i^{th}} pattern occurrence (\link{stri_locate_ith}),
or \eqn{i^{th}} text boundary (\link{stri_locate_ith_boundaries}). \cr
\item \link[=stri_join_mat]{Concatenate a character matrix row- or column-wise }. \cr
\item The \link[=strcut_loc]{strcut_-functions}.
\item Although the functions are written in R,
they have been optimized to be in somewhat the same order of speed
as the other 'stringi' functions.
}
}
\examples{

# character vector:
x <- c("3rd 1st 2nd", "5th 4th 6th")
print(x)

# detect if there are digits:
x \%s{}\% "[[:digits]]"

# cut x into matrix of individual words:
x <- strcut_brk(x, "word")

# re-order matrix using the fast \%row~\% operator:
mat <- stringi::stri_rank(as.vector(x)) |> matrix(ncol=ncol(x))
sorted <- x \%row~\% mat

# join elements of every row into a single character vector:
stri_c_mat(sorted, margin=1, sep=" ")

}
\references{
Gagolewski M., \emph{stringi: Fast and portable character string processing in R}, Journal of Statistical Software 103(2), 2022, 1–59, doi:10.18637/jss.v103.i02
}
\seealso{
\code{\link[=tinycodet_help]{tinycodet_help()}}, \code{\link[=stri_rgx]{stri_rgx()}}
}
