% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyimport_alias.R
\name{tinyimport_alias}
\alias{tinyimport_alias}
\alias{is.tinyimport_alias}
\alias{attr.import}
\title{Class tinyimport_alias}
\usage{
is.tinyimport_alias(alias)

attr.import(alias, which = NULL)
}
\arguments{
\item{alias}{the alias object as created by the \link{import_as} function.}

\item{which}{The attributes to list. If \code{NULL}, all attributes will be returned. \cr
Possibilities: "pkgs", "conflicts", and "ordered_object_names". \cr \cr}
}
\value{
For \code{is.tinyimport_alias()}: \cr
\code{TRUE} or \code{FALSE}, indicating if the object is a tiny package alias. \cr
\cr

For \code{attr.import(alias, which = NULL)}: \cr
All special attributes of the given alias object are returned as a list. \cr
\cr
For \code{attr.import(alias, which = "pkgs")}: \cr
Returns a list with 3 elements:
\itemize{
\item packages_order: a character vector of package names,
giving the packages in the order they were imported in the alias object.
\item main_package: a string giving the name of the main package.
Re-exported functions, if present, are taken together with the main package.
\item re_exports.pkgs: a character vector of package names,
giving the packages from which the re-exported functions in the main package were taken. \cr \cr
}

For \code{attr.import(alias, which = "conflicts")}: \cr
The order in which packages are imported in the alias object
(see attribute \code{pkgs$packages_order})
matters:
Functions from later named packages overwrite those from earlier named packages,
in case of conflicts. \cr
The "conflicts" attribute returns a data.frame showing exactly which functions overwrite
functions from earlier named packages, and as such "win" the conflicts. \cr
\cr
For \code{attr.import(alias, which = "ordered_object_names")}: \cr
Gives the names of the objects in the alias, in the order as they were imported. \cr
For conflicting objects, the last imported ones are used for the ordering. \cr
Note that if argument \code{re_exports} is \code{TRUE},
re-exported functions are imported when the main package is imported,
thus changing this order slightly. \cr \cr
}
\description{
The \link{import_as} function
creates an object of class \code{"tinyimport_alias"}. \cr
This help page documents its usage. \cr
\cr
To get a function from a \code{tinyimport_alias}, once can use the \code{$} operator. \cr
To use, for example, function "some_function()" from alias ".alias", use: \cr
\code{.alias$some_function()}. \cr
To "unimport" the package alias object, simply remove it from the environment it was placed in. \cr
\cr
\code{is.tinyimport_alias()} checks if an object truly is a package alias as created by the \link{import_as} function. \cr
\cr
The \code{attr.import()} function
gets one or all special attribute(s)
from an alias object returned by \link{import_as}. \cr
\cr
}
\examples{

import_as(.stri ~ stringi)
.stri$stri_join("a", "b")

import_from("stringi", import_ls("stringi", "infix"))
"a" \%s+\% "b"

attr.import(.stri)
}
