% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyoperations_dry.R
\name{tinyoperations_dry}
\alias{tinyoperations_dry}
\title{The tinyoperations "DRY" functionality}
\usage{
tinyoperations_dry()
}
\description{
"Don't Repeat Yourself", sometimes abbreviated as "DRY",
is the coding principle not to write unnecessarily repetitive code.
To help you in that effort, the \code{tinyoperations} R package
introduces a few functions:
\itemize{
\item The \link{transform_if} function
\item The \link[=\%unreal =\%]{subset_if operators and the in-place unreal modifier operator}.
\item The \link[=\%:=\%]{generalized in-place (mathematical) modification operator}.
}
}
\seealso{
\code{\link[=tinyoperations_help]{tinyoperations_help()}}
}
