% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyoperations_stringi.R
\name{tinyoperations_stringi}
\alias{tinyoperations_stringi}
\title{The tinyoperations expansion of the 'stringi' R package}
\usage{
tinyoperations_stringi()
}
\description{
The \code{tinyoperations} R package adds some functions and operators
to extend the functionality of the \code{stringi} R package:
\itemize{
\item Infix operators for \link[=\%s-\%]{string arithmetic}. \cr
\item Infix operators for \link[=\%ss\%]{string sub-setting}. \cr
\item Infix operators for \link[=\%s+ =\%]{In-place modifying string arithmetic}.
\item Infix operators for \link[=\%sget =\%]{In-place modifying string sub-setting}.
\item \link[=\%row~\%]{Infix operators for row- and column-wise re-ordering of matrices}. \cr
\item The \code{tinyoperations} package adds additional
\code{stringi} functions, namely \link{stri_locate_ith}, and
\link{stri_join_mat} (and aliases). These functions use the same naming and argument convention as the rest of
the \code{stringi} functions, thus keeping your code consistent. \cr
\item The \link[=strcut_loc]{strcut_-functions}.
\item Most \code{stringi} pattern expressions options
are available for the string-pattern-related functions, when appropriate. \cr
\item This R package has only one dependency: \code{stringi}.
No other dependencies, as to avoid \code{"dependency hell"}. \cr
\item Although the functions are written in R,
they have been aggressively optimized to be in the same order of speed
as the other \code{stringi} functions.
}

Please also have a look at the Read-Me file on the GitHub main page before using this package: \cr
\url{https://github.com/tony-aw/tinyoperations}
}
\seealso{
\code{\link[=tinyoperations_help]{tinyoperations_help()}}
}
